rename the helpers module to eth
